refactored at.RestController, allowed the Controller embedded annotation at.RestController directly
