+++ pkg/app/web/application.go
@@ -18,7 +18,7 @@ import (
 	"errors"
 	"fmt"
 	"github.com/hidevopsio/hiboot/pkg/app"
-	"github.com/hidevopsio/hiboot/pkg/app/web/at"
+	"github.com/hidevopsio/hiboot/pkg/at"
 	"github.com/hidevopsio/hiboot/pkg/log"
 	"github.com/hidevopsio/hiboot/pkg/utils/io"
 	"github.com/hidevopsio/hiboot/pkg/utils/reflector"
@@ -168,8 +168,8 @@ func (a *application) initialize(controllers ...interface{}) (err error) {
 	return
 }
 
-// Deprecated: please use app.Register() instead
 // RestController register rest controller to controllers container
+// Deprecated: please use app.Register() instead
 var RestController = app.Register
 
 // NewApplication create new web application instance and init it
