--- pkg/app/web/at/annotation_test.go
@@ -1,7 +1,0 @@
-package at
-
-import "testing"
-
-func TestDummy(t *testing.T) {
-
-}
