+++ pkg/utils/reflector/tester/foo.go
@@ -1,4 +1,5 @@
 package tester
 
+// Foo is a test interface
 type Foo interface {
 }
